[starknet] Cleanup Ragger tests
