+++ tests/test_version_cmd.py
@@ -1,14 +1,10 @@
-from application_client.command_sender import CommandSender
 from application_client.response_unpacker import unpack_get_version_response
-
 import toml
 
 # In this test we check the behavior of the device when asked to provide the app version
 def test_version(backend):
-    # Use the app interface instead of raw interface
-    client = CommandSender(backend)
     # Send the GET_VERSION instruction
-    rapdu = client.get_version()
+    rapdu = backend.exchange_raw(bytes.fromhex("5a00000000"))
     # Read version from Cargo.toml
     with open('starknet/Cargo.toml', 'r') as f:
         config = toml.load(f)
