+++ tests/usage.md
@@ -15,12 +15,13 @@ sudo apt-get update && sudo apt-get install qemu-user-static
 ### Compile the application
 
 The application to test must be compiled for all required devices.
-You can use for this the container `ghcr.io/ledgerhq/ledger-app-builder/ledger-app-builder-lite`:
+You can use for this the container `ghcr.io/ledgerhq/ledger-app-builder/ledger-app-builder`:
 ```
-docker pull ghcr.io/ledgerhq/ledger-app-builder/ledger-app-builder-lite:latest
-cd <your app repository>                        # replace <appname> with the name of your app, (eg boilerplate)
-docker run --user "$(id -u)":"$(id -g)" --rm -ti -v "$(realpath .):/app" --privileged -v "/dev/bus/usb:/dev/bus/usb" ledger-app-builder-lite:latest
-make clean && make BOLOS_SDK=$<device>_SDK      # replace <device> with one of [NANOS, NANOX, NANOSP, STAX]
+docker pull ghcr.io/ledgerhq/ledger-app-builder/ledger-app-builder:latest
+cd app-starknet 
+docker run --user "$(id -u)":"$(id -g)" --rm -ti -v "$(realpath .):/app" --privileged -v "/dev/bus/usb:/dev/bus/usb" ledger-app-builder:latest
+cd /app/starknet
+cargo ledger build <device>      # replace <device> with one of [nanosplus, nanox, stax, flex]
 exit
 ```
 
@@ -28,7 +29,8 @@ exit
 
 You can use the following command to get your first experience with Ragger and Speculos
 ```
-pytest -v --tb=short --device nanox --display
+pip install -r tests/requirement.txt
+pytest tests -v --tb=short --device stax
 ```
 Or you can refer to the section `Available pytest options` to configure the options you want to use
 
@@ -36,19 +38,16 @@ Or you can refer to the section `Available pytest options` to configure the opti
 ### Run a simple test using a real device
 
 The application to test must be loaded and started on a Ledger device plugged in USB.
-You can use for this the container `ghcr.io/ledgerhq/ledger-app-builder/ledger-app-builder-lite`:
+You can use `ledgerwallet` to sideload the app on a Ledger device (e.g. Stax):
 ```
-docker pull ghcr.io/ledgerhq/ledger-app-builder/ledger-app-builder-lite:latest
-cd app-<appname>/                                   # replace <appname> with the name of your app, (eg boilerplate)
-docker run --user "$(id -u)":"$(id -g)" --rm -ti -v "$(realpath .):/app" --privileged -v "/dev/bus/usb:/dev/bus/usb" ledger-app-builder-lite:latest
-make clean && make BOLOS_SDK=$<device>_SDK load     # replace <device> with one of [NANOS, NANOX, NANOSP, STAX]
-exit
+pip install ledgerwallet
+ledgerctl install -f target/stax/release/app_stax.json
 ```
 
 You can use the following command to get your first experience with Ragger and Ledgerwallet on a NANOX.
 Make sure that the device is plugged, unlocked, and that the tested application is open.
 ```
-pytest -v --tb=short --device nanox --backend ledgerwallet
+pytest -v --tb=short --device stax --backend ledgerwallet
 ```
 Or you can refer to the section `Available pytest options` to configure the options you want to use
